requirements.txt
discord.py
discord-py-interactions